backend/requirements.txt
# ── FRAMEWORK ──
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── DATABASE ──
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# ── AUTH ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# ── AI ──
google-generativeai==0.5.4

# ── WEBSOCKET ──
websockets==12.0

# ── PAYMENTS ──
httpx==0.27.0
requests==2.31.0

# ── AFRICA'S TALKING (USSD + SMS) ──
africastalking==1.2.5

# ── SCHEDULING ──
apscheduler==3.10.4

# ── ENCRYPTION ──
cryptography==42.0.5

# ── TIMEZONE ──
pytz==2024.1

# ── UTILITIES ──
python-dotenv==1.0.1

# ── TESTING ──
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0

